inbox: extract routes into own classes
