+++ slushy-app/src/main/java/net/kemitix/slushy/app/inbox/InboxSendEmailRoute.java
@@ -1,0 +1,18 @@
+package net.kemitix.slushy.app.inbox;
+
+import org.apache.camel.builder.RouteBuilder;
+
+import javax.enterprise.context.ApplicationScoped;
+
+@ApplicationScoped
+public class InboxSendEmailRoute
+        extends RouteBuilder {
+    @Override
+    public void configure() {
+        from("direct:Slushy.Inbox.SendEmailConfirmation")
+                .routeId("Slushy.Inbox.SendEmailConfirmation")
+                .setHeader("SlushyEmailTemplate").constant("inbox")
+                .to("direct:Slushy.SendEmail")
+        ;
+    }
+}
