+++ slushy-app/src/main/java/net/kemitix/slushy/app/inbox/InboxTimerRoute.java
@@ -1,0 +1,39 @@
+package net.kemitix.slushy.app.inbox;
+
+import net.kemitix.slushy.app.IsRequiredAge;
+import net.kemitix.slushy.app.OnException;
+import net.kemitix.slushy.app.trello.TrelloBoard;
+import org.apache.camel.builder.RouteBuilder;
+
+import javax.enterprise.context.ApplicationScoped;
+import javax.inject.Inject;
+
+import static org.apache.camel.builder.Builder.bean;
+
+@ApplicationScoped
+public class InboxTimerRoute
+        extends RouteBuilder {
+
+    @Inject InboxConfig inboxConfig;
+    @Inject TrelloBoard trelloBoard;
+    @Inject IsRequiredAge isRequiredAge;
+
+    @Override
+    public void configure() {
+        OnException.retry(this, inboxConfig);
+
+        fromF("timer:inbox?period=%s", inboxConfig.getScanPeriod())
+                .routeId("Slushy.Inbox")
+
+                .setBody(exchange -> trelloBoard.getListCards(inboxConfig.getSourceList()))
+                .split(body())
+
+                .setHeader("SlushyRequiredAge", inboxConfig::getRequiredAgeHours)
+                .filter(bean(isRequiredAge))
+
+                .setHeader("SlushyRoutingSlip", inboxConfig::getRoutingSlip)
+                .routingSlip(header("SlushyRoutingSlip"))
+        ;
+    }
+
+}
