+++ slushy-app/src/main/java/net/kemitix/slushy/app/inbox/ReformatCardRoute.java
@@ -1,0 +1,25 @@
+package net.kemitix.slushy.app.inbox;
+
+import org.apache.camel.builder.RouteBuilder;
+
+import javax.enterprise.context.ApplicationScoped;
+import javax.inject.Inject;
+
+@ApplicationScoped
+public class ReformatCardRoute
+        extends RouteBuilder {
+    private final ReformatCard reformatCard;
+
+    @Inject
+    public ReformatCardRoute(ReformatCard reformatCard) {
+        this.reformatCard = reformatCard;
+    }
+
+    @Override
+    public void configure() {
+        from("direct:Slushy.ReformatCard")
+                .routeId("Slushy.ReformatCard")
+                .bean(reformatCard)
+        ;
+    }
+}
